# Some things to make the compiler faster when built in debug