--- a/bootstrap/stage0/repl_backend__default.cpp
+++ b/bootstrap/stage0/repl_backend__default.cpp
@@ -13,7 +13,7 @@ builder.appendff("prompt: \"{}\"", prompt);
 builder.append(")"sv);return builder.to_string(); }
 ErrorOr<repl_backend__default::Editor> repl_backend__default::Editor::create(ByteString const prompt,Function<ErrorOr<void>(repl_backend__default::Editor&)> const& syntax_highlight_handler) {
 {
-FILE* std_in = fopen((((ByteString::must_from_utf8("/dev/stdin"sv))).characters()),(((ByteString::must_from_utf8("r"sv))).characters()));
+FILE* std_in = fopen((((ByteString::from_utf8_without_validation("/dev/stdin"sv))).characters()),(((ByteString::from_utf8_without_validation("r"sv))).characters()));
 if ((std_in == jakt__platform__utility::null<FILE*>())){
 warnln((StringView::from_string_literal("Could not open /dev/stdin for reading"sv)));
 return Error::from_errno(static_cast<i32>(42));