… bootstrap/stage0/codegen.cpp
… bootstrap/stage0/formatter.cpp
… bootstrap/stage0/ide.cpp
… bootstrap/stage0/interpreter.cpp
… bootstrap/stage0/lexer.cpp
… bootstrap/stage0/main.cpp
… bootstrap/stage0/parser.cpp
… bootstrap/stage0/repl.cpp
… bootstrap/stage0/runtime/jaktlib/prelude/string.jakt
… bootstrap/stage0/typechecker.cpp
… bootstrap/stage0/types.cpp
… bootstrap/stage0/utility.cpp
… runtime/jaktlib/prelude/string.jakt
… selfhost/types.jakt