Together with https://github.com/SerenityOS/serenity/pull/22377 these changes make compiling the compiler with itself in debug mode ~44% faster.

```
$ time build/bin/jakt_before -S selfhost/main.jakt 
real    0m23,247s
user    0m22,917s
sys    0m0,212s

$ time build/bin/jakt_after -S selfhost/main.jakt 
real    0m16,072s
user    0m15,875s
sys    0m0,192s
```